decoder: harden frame arbitration and FEC state
Failed SSDV candidates use only image and packet IDs as their arbitration
key.  Different spacecraft can therefore collapse into one candidate when
their identifiers overlap.  Recorded playback also leaves TCP and ZMQ
transports enabled.

Include the spacecraft header in candidate identity and disable network
transports during replay.  Reset SSDV state whenever the source, geometry,
quality, or first packet changes.

Correct the OpenHoshimi AGC eviction order.  Initialize CCSDS state and
use unsigned shifts for Viterbi decisions and SSDV fields.  Add tests for
concurrent first-wins delivery, normalized PDU output, and SSDV session
changes.

Fixes: f7b38ac68337 ("upload: restore telemetry frame publication")
Fixes: c7d059e01ca3 ("feat(dsp): add OpenHoshimi parallel decoder")
Fixes: fe0126898086 ("feat(dsp): SSDV image reception; scroll waterfall during fast replay")

+++ libs/demod/frame_monitor.h
@@ -49,7 +49,7 @@ private:
     void handleLocalCandidate(const pmt::pmt_t& message,
                               CandidatePriority priority);
     void candidateWorker();
-    static std::uint32_t candidateKey(const std::vector<std::uint8_t>& payload);
+    static std::uint64_t candidateKey(const std::vector<std::uint8_t>& payload);
     static std::string describePdu(const pmt::pmt_t& message);
 
     Callback callback_;
@@ -78,7 +78,7 @@ private:
     };
     std::mutex candidate_mutex_;
     std::condition_variable candidate_cv_;
-    std::map<std::uint32_t, PendingCandidate> pending_candidates_;
+    std::map<std::uint64_t, PendingCandidate> pending_candidates_;
     bool stop_candidate_worker_ = false;
     std::thread candidate_worker_;
 };
